+++ src/main/java/no/novari/fint/model/resource/FintLinks.java
@@ -1,4 +1,4 @@
-package no.novari.model.resource;
+package no.novari.fint.model.resource;
 
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
