+++ src/main/java/no/novari/fint/model/resource/FintResource.java
@@ -1,0 +1,6 @@
+package no.novari.fint.model.resource;
+
+import no.novari.fint.model.FintMainObject;
+
+public interface FintResource extends FintLinks, FintMainObject {
+}
