--- src/main/java/no/novari/model/resource/FintResource.java
@@ -1,6 +1,0 @@
-package no.novari.model.resource;
-
-import no.novari.model.FintMainObject;
-
-public interface FintResource extends FintLinks, FintMainObject {
-}
