+++ src/test/groovy/no/novari/fint/model/resource/AbstractCollectionResourcesSpec.groovy
@@ -1,8 +1,8 @@
-package no.novari.model.resource
+package no.novari.fint.model.resource
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import no.novari.model.resource.testutils.PersonResource
-import no.novari.model.resource.testutils.PersonResources
+import no.novari.fint.model.resource.testutils.PersonResource
+import no.novari.fint.model.resource.testutils.PersonResources
 import spock.lang.Specification
 
 class AbstractCollectionResourcesSpec extends Specification {
