+++ src/test/groovy/no/novari/fint/model/resource/FintLinksSpec.groovy
@@ -1,6 +1,7 @@
-package no.novari.model.resource
+package no.novari.fint.model.resource
 
-import no.novari.model.resource.testutils.PersonResource
+
+import no.novari.fint.model.resource.testutils.PersonResource
 import spock.lang.Specification
 
 class FintLinksSpec extends Specification {
