+++ src/test/groovy/no/novari/fint/model/resource/testutils/PersonResources.java
@@ -1,8 +1,8 @@
-package no.novari.model.resource.testutils;
+package no.novari.fint.model.resource.testutils;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import lombok.NoArgsConstructor;
-import no.novari.model.resource.AbstractCollectionResources;
+import no.novari.fint.model.resource.AbstractCollectionResources;
 
 import java.util.Collection;
 import java.util.List;
@@ -16,7 +16,7 @@ public class PersonResources extends AbstractCollectionResources<PersonResource>
     @Deprecated
     @Override
     public TypeReference<List<PersonResource>> getTypeReference() {
-        return new TypeReference<>() {
+        return new TypeReference<List<PersonResource>>() {
         };
     }
 }
