+++ src/test/groovy/no/novari/fint/model/testutils/Person.java
@@ -1,0 +1,4 @@
+package no.novari.fint.model.testutils;
+
+public class Person {
+}
