--- src/test/groovy/no/novari/model/testutils/Person.java
@@ -1,4 +1,0 @@
-package no.novari.model.testutils;
-
-public class Person {
-}
